Add version-specific implementations for methods with argument-passing.
